refactor(cli): rename tools/linear.mjs to tools/ticket.mjs with a deprecated shim (WM-1026) (#813)
